feat: Included error 404 Http Not Found

+++ src/controllers/delete_person_controller.py
@@ -2,6 +2,7 @@ from typing import Type, Dict
 from src.models.person_repository import PersonRepository
 from src.models.tuples import PersonsTuple
 from .interface.delete_person_interface import DeletePersonInterface
+from src.errors.http_not_found import HttpNotFound
 
 
 class DeletePersonController(DeletePersonInterface):
@@ -16,14 +17,17 @@ class DeletePersonController(DeletePersonInterface):
         :return - Dictionary with informations of the process
         """
 
-        try:
-            self.__convert_validate(name)
-            new_person = self.person_repository.delete_person(
+        self.__convert_validate(name)
+        person = self.person_repository.select_person(
+            name
+        )
+        if person:
+            deleted_person = self.person_repository.delete_person(
                 name
             )
-            return {"success": True, "person_registry": new_person}
-        except Exception as exception:
-            return {"success": False, "error": str(exception)}
+            return {"success": True, "person_registry": deleted_person}
+        else:
+            raise HttpNotFound(f"{name} not found")
 
     def __convert_validate(self, name: str):
         validate_entry = (
